+++ backend/src/main/java/com/xfactor/openlibrary/repository/AdminRepository.java
@@ -2,6 +2,6 @@ package com.xfactor.openlibrary.repository;
 import com.xfactor.openlibrary.domain.admin;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface AdminRepository extends JpaRepository<admin, Integer>{
+public interface AdminRepository extends JpaRepository<admin, Long>{
     
 }
