+++ backend/src/main/java/com/xfactor/openlibrary/repository/PublisherRepository.java
@@ -4,6 +4,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import com.xfactor.openlibrary.domain.Publisher;
 
-public interface PublisherRepository extends JpaRepository<Publisher, Integer> {
+public interface PublisherRepository extends JpaRepository<Publisher, Long> {
     
 }
